Add Business Source License 1.1
This commit adds the Business Source License 1.1 (BSL) to the repository.

The BSL license allows non-commercial use of the code and requires commercial users to obtain prior written permission from the Licensor. On 2026-06-15, the code will transition to the GNU General Public License v2.0 or any later version.

- License ID: Steven-Sluder License
- Licensor: Steven Sluder
- Effective Date: 2024-06-15
- Change Date: 2026-06-15
- Change License: GNU General Public License v2.0 or any later version

For more details, see the LICENSE file.